refactor run-policy for custom policies
